sample/src/main/kotlin/io/sweers/moshisealed/sample/Message.kt
package io.sweers.moshisealed.sample

import com.squareup.moshi.JsonClass
import io.sweers.moshisealed.annotations.DefaultObject
import io.sweers.moshisealed.annotations.TypeLabel

@JsonClass(generateAdapter = true, generator = "sealed:type")
sealed class Message {

  @TypeLabel("success")
  @JsonClass(generateAdapter = true)
  data class Success(val value: String) : Message()

  @TypeLabel("error")
  @JsonClass(generateAdapter = true)
  data class Error(val error_logs: Map<String, Any>) : Message()

  @DefaultObject
  object Unknown : Message()
}
